Fix Step Cloner get_children
