+++ testing/test_apply.py
@@ -1,42 +1,28 @@
-import numpy as np
-
+from neuraxle.base import Identity
 from neuraxle.hyperparams.space import HyperparameterSamples
 from neuraxle.pipeline import Pipeline
-from neuraxle.steps.misc import TapeCallbackFunction
 from neuraxle.steps.numpy import MultiplyByN
 from neuraxle.steps.output_handlers import OutputTransformerWrapper
-
-DATA_INPUTS = np.array([1])
-EXPECTED_OUTPUTS = np.array([1])
-
-tape_transform = TapeCallbackFunction()
-tape_fit = TapeCallbackFunction()
-
-
-def test_apply_on_pipeline_should_call_method_on_each_steps():
-    pipeline = Pipeline([MultiplyByN(1), MultiplyByN(1)])
-
-    pipeline.apply('set_hyperparams', hyperparams=HyperparameterSamples({'multiply_by': 2}))
-
-    assert pipeline.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN1'].get_hyperparams()['multiply_by'] == 2
 
 
 def test_apply_on_pipeline_with_positional_argument_should_call_method_on_each_steps():
     pipeline = Pipeline([MultiplyByN(1), MultiplyByN(1)])
 
-    pipeline.apply('set_hyperparams', hyperparams=HyperparameterSamples({'multiply_by': 2}))
+    pipeline.apply('set_hyperparams', hyperparams=HyperparameterSamples({
+        'multiply_by': 2,
+        'MultiplyByN__multiply_by': 3,
+        'MultiplyByN1__multiply_by': 4
+    }))
 
     assert pipeline.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN1'].get_hyperparams()['multiply_by'] == 2
+    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 3
+    assert pipeline['MultiplyByN1'].get_hyperparams()['multiply_by'] == 4
 
 
 def test_apply_method_on_pipeline_should_call_method_on_each_steps():
     pipeline = Pipeline([MultiplyByN(1), MultiplyByN(1)])
 
-    pipeline.apply_method(
+    pipeline.apply(
         lambda step: step.set_hyperparams(HyperparameterSamples({'multiply_by': 2}))
     )
 
@@ -45,59 +31,98 @@ def test_apply_method_on_pipeline_should_call_method_on_each_steps():
     assert pipeline['MultiplyByN1'].get_hyperparams()['multiply_by'] == 2
 
 
-def test_apply_method_on_pipeline_with_positional_argument_should_call_method_on_each_steps():
-    pipeline = Pipeline([MultiplyByN(1), MultiplyByN(1)])
-
-    pipeline.apply_method(
-        lambda step, hyperparams: step.set_hyperparams(hyperparams),
-        hyperparams=HyperparameterSamples({'multiply_by': 2})
-    )
-
-    assert pipeline.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN1'].get_hyperparams()['multiply_by'] == 2
-
-
-def test_apply_on_pipeline_with_meta_step_should_call_method_on_each_steps():
+def test_apply_on_pipeline_with_meta_step_and_positional_argument():
     pipeline = Pipeline([OutputTransformerWrapper(MultiplyByN(1)), MultiplyByN(1)])
 
-    pipeline.apply('set_hyperparams', hyperparams=HyperparameterSamples({'multiply_by': 2}))
+    pipeline.apply('set_hyperparams', hyperparams=HyperparameterSamples({
+        'multiply_by': 2,
+        'OutputTransformerWrapper__multiply_by': 3,
+        'OutputTransformerWrapper__MultiplyByN__multiply_by': 4,
+        'MultiplyByN__multiply_by': 5
+    }))
 
     assert pipeline.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['OutputTransformerWrapper'].wrapped.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 2
-
-
-def test_apply_on_pipeline_with_meta_step_and_positional_argument_should_call_method_on_each_steps():
-    pipeline = Pipeline([OutputTransformerWrapper(MultiplyByN(1)), MultiplyByN(1)])
-
-    pipeline.apply('set_hyperparams', hyperparams=HyperparameterSamples({'multiply_by': 2}))
-
-    assert pipeline.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['OutputTransformerWrapper'].wrapped.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 2
+    assert pipeline['OutputTransformerWrapper'].get_hyperparams()['multiply_by'] == 3
+    assert pipeline['OutputTransformerWrapper'].wrapped.get_hyperparams()['multiply_by'] == 4
+    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 5
 
 
 def test_apply_method_on_pipeline_with_meta_step_should_call_method_on_each_steps():
     pipeline = Pipeline([OutputTransformerWrapper(MultiplyByN(1)), MultiplyByN(1)])
 
-    pipeline.apply_method(
+    pipeline.apply(
         lambda step: step.set_hyperparams(HyperparameterSamples({'multiply_by': 2}))
     )
 
     assert pipeline.get_hyperparams()['multiply_by'] == 2
+    assert pipeline['OutputTransformerWrapper'].get_hyperparams()['multiply_by'] == 2
     assert pipeline['OutputTransformerWrapper'].wrapped.get_hyperparams()['multiply_by'] == 2
     assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 2
 
 
-def test_apply_method_on_pipeline_with_meta_step_and_positional_argument_should_call_method_on_each_steps():
-    pipeline = Pipeline([OutputTransformerWrapper(MultiplyByN(1)), MultiplyByN(1)])
+def test_has_children_mixin_apply_should_apply_method_to_direct_childrends():
+    p = Pipeline([
+        ('a', Identity()),
+        ('b', Identity()),
+        Pipeline([
+            ('c', Identity()),
+            ('d', Identity())
+        ]),
+    ])
 
-    pipeline.apply_method(
-        lambda step, hyperparams: step.set_hyperparams(hyperparams),
-        hyperparams=HyperparameterSamples({'multiply_by': 2})
-    )
+    p.apply('set_hyperparams', ra=None, hyperparams=HyperparameterSamples({
+        'a__hp': 0,
+        'b__hp': 1,
+        'Pipeline__hp': 2
+    }))
 
-    assert pipeline.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['OutputTransformerWrapper'].wrapped.get_hyperparams()['multiply_by'] == 2
-    assert pipeline['MultiplyByN'].get_hyperparams()['multiply_by'] == 2
+    assert p['a'].hyperparams.to_flat_as_dict_primitive()['hp'] == 0
+    assert p['b'].hyperparams.to_flat_as_dict_primitive()['hp'] == 1
+    assert p['Pipeline'].hyperparams.to_flat_as_dict_primitive()['hp'] == 2
+
+
+def test_has_children_mixin_apply_should_apply_method_to_recursive_childrends():
+    p = Pipeline([
+        ('a', Identity()),
+        ('b', Identity()),
+        Pipeline([
+            ('c', Identity()),
+            ('d', Identity())
+        ]),
+    ])
+
+    p.apply('set_hyperparams', ra=None, hyperparams=HyperparameterSamples({
+        'Pipeline__c__hp': 3,
+        'Pipeline__d__hp': 4
+    }))
+
+    assert p['Pipeline']['c'].hyperparams.to_flat_as_dict_primitive()['hp'] == 3
+    assert p['Pipeline']['d'].hyperparams.to_flat_as_dict_primitive()['hp'] == 4
+
+
+def test_has_children_mixin_apply_should_return_recursive_dict_to_direct_childrends():
+    p = Pipeline([
+        ('a', Identity().set_hyperparams(HyperparameterSamples({'hp': 0}))),
+        ('b', Identity().set_hyperparams(HyperparameterSamples({'hp': 1})))
+    ])
+
+    results = p.apply('get_hyperparams', ra=None)
+
+    assert results.to_flat_as_dict_primitive()['a__hp'] == 0
+    assert results.to_flat_as_dict_primitive()['b__hp'] == 1
+
+
+def test_has_children_mixin_apply_should_return_recursive_dict_to_recursive_childrends():
+    p = Pipeline([
+        Pipeline([
+            ('c', Identity().set_hyperparams(HyperparameterSamples({'hp': 3}))),
+            ('d', Identity().set_hyperparams(HyperparameterSamples({'hp': 4})))
+        ]).set_hyperparams(HyperparameterSamples({'hp': 2})),
+    ])
+
+    results = p.apply('get_hyperparams', ra=None)
+    results = results.to_flat_as_dict_primitive()
+
+    assert results['Pipeline__hp'] == 2
+    assert results['Pipeline__c__hp'] == 3
+    assert results['Pipeline__d__hp'] == 4
